backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
emails>=0.6.0,<0.7.0
psycopg2-binary>=2.9.1,<2.10.0
alembic>=1.7.0,<1.8.0
email-validator>=1.1.2,<1.2.0
jinja2>=3.0.0,<3.1.0 